+++ src/plugins/dummyPlugin.h
@@ -37,7 +37,7 @@ public:
     DummyPlugin(Profile*, ContextTracker*);
     ~DummyPlugin();
 
-    virtual Prediction predict(const int) const;
+    virtual Prediction predict(const size_t) const;
 
     virtual void learn();
     virtual void extract();
